Add decimal implemenation and port indicator usages

+++ indicator.go
@@ -1,5 +1,5 @@
 package talib4g
 
 type Indicator interface {
-	Calculate(int) float64
+	Calculate(int) Decimal
 }
